Fix some C++Test violations.
Change c-style casts to c++ style casts.

(From Sierra)

+++ packages/stk/stk_util/stk_util/environment/memory_util.cpp
@@ -154,7 +154,7 @@ void get_processor_count(std::vector<int> &procinfo)
     {
       proc_cpuinfo.close();
 
-      if(procinfo.size() == (size_t) 0)
+      if(procinfo.size() == 0u)
 	procinfo.push_back(1);  
 
       return;
@@ -181,7 +181,7 @@ void get_processor_count(std::vector<int> &procinfo)
 
     if(have_processor)
     {
-      if((size_t) proc == procinfo.size())
+      if(static_cast<size_t>(proc) == procinfo.size())
 	procinfo.push_back(cores);
 
       have_processor = false;
